requirements.txt
smartapi-python
pandas
numpy
scipy
streamlit
python-telegram-bot
pyotp
python-dotenv
plotly
streamlit-autorefresh
